+++ addons/mx/cfdi/bill_invoice.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/invopop/gobl/bill"
 	"github.com/invopop/gobl/cal"
+	"github.com/invopop/gobl/currency"
 	"github.com/invopop/gobl/head"
 	"github.com/invopop/gobl/num"
 	"github.com/invopop/gobl/org"
@@ -37,6 +38,7 @@ func normalizeInvoiceIssueDateAndTime(inv *bill.Invoice) {
 
 func billInvoiceRules() *rules.Set {
 	return rules.For(new(bill.Invoice),
+		rules.Assert("27", "invoice must be in MXN or provide exchange rate for conversion", currency.CanConvertTo(currency.MXN)),
 		// Tax extensions
 		rules.Field("tax",
 			rules.Field("ext",